+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-# 0.1.0
+# 1.0.0
++ Add login method
++ Add CircleCI tests
++ Add delete tag call
+
+## 0.1.0
 + First commit